Cleanup cursor logic.
