+++ src/Murder.Editor/Core/Cursor/CursorStyle.cs
@@ -1,0 +1,9 @@
+﻿namespace Murder.Editor.Core;
+
+public enum CursorStyle
+{
+    Normal,
+    Hand,
+    Point,
+    Eye
+}
