--- src/Murder.Editor/Core/CursorStyle.cs
@@ -1,10 +1,0 @@
-﻿namespace Murder.Editor.EditorCore
-{
-    public enum CursorStyle
-    {
-        Normal,
-        Hand,
-        Point,
-        Eye
-    }
-}
