+++ src/Murder.Editor/Systems/Debug/DebugShowCameraBoundsSystem.cs
@@ -4,7 +4,7 @@ using Murder.Core.Geometry;
 using Murder.Core.Graphics;
 using Murder.Core.Input;
 using Murder.Editor.Components;
-using Murder.Editor.EditorCore;
+using Murder.Editor.Core;
 using Murder.Editor.Utilities;
 using Murder.Services;
 using System.Numerics;
