+++ src/Murder.Editor/Systems/EditorStartOnCursorSystem.cs
@@ -8,7 +8,7 @@ using Murder.Core.Graphics;
 using Murder.Core.Input;
 using Murder.Editor.Attributes;
 using Murder.Editor.Components;
-using Murder.Editor.EditorCore;
+using Murder.Editor.Core;
 using Murder.Editor.Services;
 using Murder.Editor.Utilities;
 using Murder.Services;
